+++ src/github.com/nlamirault/abraracoursix/api/v1/webservice.go
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package v1
 
 import (
-	"fmt"
+	// "fmt"
 	"net/http"
 
 	log "github.com/Sirupsen/logrus"
@@ -24,9 +24,9 @@ import (
 	"github.com/nlamirault/abraracoursix/storage"
 )
 
-// var (
-// 	webservice *WebService
-// )
+type URL struct {
+	URL string `json:"url" binding:"required"`
+}
 
 // WebService represents the Restful API
 type WebService struct {
@@ -35,6 +35,7 @@ type WebService struct {
 
 // NewWebService creates a new WebService instance
 func NewWebService(store storage.Storage) *WebService {
+	log.Debugf("Creates webservice with backend : %v", store)
 	return &WebService{Store: store}
 }
 
@@ -47,36 +48,4 @@ func (ws *WebService) Help(c *gin.Context) {
 // DisplayAPIVersion sends the API version in JSON format
 func (ws *WebService) DisplayAPIVersion(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{"version": "1"})
-}
-
-// URLShow send the url store using the key
-func (ws *WebService) URLShow(c *gin.Context) {
-	key := c.Param("url")
-	log.Info("Retrieve URL using key: ", key)
-	data, err := ws.Store.Get([]byte(key))
-	if err != nil {
-		str := fmt.Sprintf("Error retrieving URL with key %s", key)
-		c.JSON(http.StatusInternalServerError, gin.H{"Error": str})
-		return
-	}
-	if data == nil {
-		c.JSON(http.StatusNotFound, gin.H{"Unknown key": key})
-		return
-	}
-	url := string(data)
-	log.Info("Find URL : ", url)
-	c.JSON(http.StatusOK, gin.H{"URL": url})
-}
-
-// URLCreate store a long URL using a key
-func (ws *WebService) URLCreate(c *gin.Context) {
-	url := c.Param("url")
-	key := "aaa"
-	err := ws.Store.Put([]byte(key), []byte(url))
-	if err != nil {
-		str := fmt.Sprintf("Can't store URL %s", url)
-		c.JSON(http.StatusNotFound, gin.H{"Error": str})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"ShortURL": key})
 }