+++ ChangeLog.md
@@ -3,3 +3,6 @@ Abraracoursix ChangeLog
 
 # Version 0.1.0 (IN PROGRESS)
 
+- API to store and retrieve URL
+- Add LevelDB backend
+- Add BoltDB backend